Refactor: Record -> RunningRecord
