+++ relayRun/src/main/java/com/example/relayRun/record/controller/RunningRecordController.java
@@ -1,5 +1,6 @@
 package com.example.relayRun.record.controller;
 
+import com.example.relayRun.record.dto.GetRecordByIdxRes;
 import com.example.relayRun.record.dto.PostRunningInitReq;
 import com.example.relayRun.record.dto.PostRunningInitRes;
 import com.example.relayRun.record.service.RunningRecordService;
@@ -7,10 +8,7 @@ import com.example.relayRun.util.BaseException;
 import com.example.relayRun.util.BaseResponse;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/record")
@@ -21,6 +19,8 @@ public class RunningRecordController {
     public RunningRecordController(RunningRecordService runningRecordService){
         this.runningRecordService = runningRecordService;
     }
+
+    // 달리기 시작
     @PostMapping("/start")
     @ApiOperation(value="달리기 시작", notes="profileIdx, 속한 clubIdx 요청")
     public BaseResponse<PostRunningInitRes> startRunning(@RequestBody PostRunningInitReq runningInitReq) {
@@ -29,6 +29,18 @@ public class RunningRecordController {
             return new BaseResponse<>(result);
         }catch(BaseException e) {
             return new BaseResponse<>(e.getStatus());
+        }
+    }
+
+    // 기록 세부 조회
+    @ApiOperation(value="기록 idx로 조회", notes="path variable에 조회할 기록의 idx를 입력해주세요")
+    @GetMapping("/{idx}")
+    public BaseResponse<GetRecordByIdxRes> getRecordByIdx(@PathVariable("idx") Long idx) {
+        try {
+            GetRecordByIdxRes rec = runningRecordService.getRecordByIdx(idx);
+            return new BaseResponse<>(rec);
+        } catch (BaseException e) {
+            return new BaseResponse(e.getStatus());
         }
     }
 }
