+++ relayRun/src/main/java/com/example/relayRun/record/dto/GetRecordByIdxRes.java
@@ -1,6 +1,7 @@
 package com.example.relayRun.record.dto;
 
 import com.example.relayRun.record.entity.LocationEntity;
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -11,11 +12,16 @@ import java.util.List;
 @Builder
 public class GetRecordByIdxRes {
     private Long recordIdx;
-//    private String teamName;
+
+    @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
     private LocalDateTime date;
+
     private Float time;
+
     private Float distance;
+
     private Float pace;
+
     private String goalStatus;
 
     private List<GetLocationRes> locationList;
