+++ relayRun/src/main/java/com/example/relayRun/record/repository/RunningRecordRepository.java
@@ -3,5 +3,8 @@ package com.example.relayRun.record.repository;
 import com.example.relayRun.record.entity.RunningRecordEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface RunningRecordRepository extends JpaRepository<RunningRecordEntity, Long> {
+    Optional<RunningRecordEntity> findByRunningRecordIdxAndStatus(Long idx, String status);
 }
