+++ relayRun/src/main/java/com/example/relayRun/record/service/RunningRecordService.java
@@ -2,8 +2,11 @@ package com.example.relayRun.record.service;
 
 import com.example.relayRun.club.entity.MemberStatusEntity;
 import com.example.relayRun.club.repository.MemberStatusRepository;
+import com.example.relayRun.record.dto.GetLocationRes;
+import com.example.relayRun.record.dto.GetRecordByIdxRes;
 import com.example.relayRun.record.dto.PostRunningInitReq;
 import com.example.relayRun.record.dto.PostRunningInitRes;
+import com.example.relayRun.record.entity.LocationEntity;
 import com.example.relayRun.record.entity.RunningRecordEntity;
 import com.example.relayRun.record.repository.RunningRecordRepository;
 import com.example.relayRun.util.BaseException;
@@ -11,6 +14,8 @@ import com.example.relayRun.util.BaseResponseStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -24,6 +29,13 @@ public class RunningRecordService {
         this.memberStatusRepository = memberStatusRepository;
     }
 
+
+    /**
+     * 달리기 시작 POST
+     * @param runningInitReq
+     * @return
+     * @throws BaseException
+     */
     public PostRunningInitRes startRunning(PostRunningInitReq runningInitReq) throws BaseException {
         Optional<MemberStatusEntity> optionalMemberStatus = memberStatusRepository.findByUserProfileIdx_UserProfileIdxAndApplyStatusIs(
                 runningInitReq.getProfileIdx(),
@@ -41,5 +53,54 @@ public class RunningRecordService {
         PostRunningInitRes result = new PostRunningInitRes();
         result.setRunningRecordIdx(recordEntity.getRunningRecordIdx());
         return result;
+    }
+
+
+    /**
+     * 기록 세부 조회 GET
+     * @param idx
+     * @return
+     * @throws BaseException
+     */
+    public GetRecordByIdxRes getRecordByIdx(Long idx) throws BaseException {
+        try {
+            Optional<RunningRecordEntity> record = runningRecordRepository.findByRunningRecordIdxAndStatus(idx, "active");
+            if (record.isEmpty()) {
+                throw new Exception("RECORD_UNAVAILABLE");
+            }
+
+//            List<GetLocationRes> locationList = locationRepository.findByRecordIdx_RunningRecordIdx(idx);
+            List<LocationEntity> getLocations = record.get().getLocations();
+
+            List<GetLocationRes> locationList = new ArrayList<>();
+            for (LocationEntity location : getLocations) {
+                locationList.add(
+                        GetLocationRes.builder()
+                                .time(location.getTime())
+                                .longitude((float) location.getPosition().getX())
+                                .latitude((float) location.getPosition().getY())
+                                .status(location.getStatus())
+                                .build()
+                );
+            }
+
+            return GetRecordByIdxRes.builder()
+                    .recordIdx(idx)
+                    .date(record.get().getCreatedAt())
+                    .time(record.get().getTime())
+                    .distance(record.get().getDistance())
+                    .pace(record.get().getPace())
+                    .goalStatus(record.get().getGoalStatus())
+                    .locationList(locationList)
+                    .build();
+
+        } catch (Exception e) {
+            if (e.getMessage().equals("RECORD_UNAVAILABLE")) {
+                throw new BaseException(BaseResponseStatus.RECORD_UNAVAILABLE);
+            } else {
+                System.out.println("e = " + e);
+                throw new BaseException(BaseResponseStatus.DATABASE_ERROR);
+            }
+        }
     }
 }
